Relocate mutations and remaining invoice query test to graphql crate
