+++ graphql/invoice/src/mutations/outbound_shipment/insert.rs
@@ -102,3 +102,227 @@ fn map_error(error: ServiceError) -> Result<InsertErrorInterface> {
 
     Err(graphql_error.extend())
 }
+
+#[cfg(test)]
+mod graphql {
+    use async_graphql::EmptyMutation;
+    use chrono::{DateTime, Utc};
+    use graphql_core::test_helpers::setup_graphl_test;
+    use graphql_core::{
+        assert_graphql_query, assert_standard_graphql_error, get_invoice_lines_inline,
+    };
+    use repository::mock::{
+        mock_name_linked_to_store, mock_name_not_linked_to_store,
+        mock_outbound_shipment_number_store_a, mock_store_linked_to_name,
+    };
+    use repository::{mock::MockDataInserts, InvoiceFilter, InvoiceQueryRepository};
+    use repository::{EqualFilter, InvoiceRepository};
+    use serde_json::json;
+    use util::uuid::uuid;
+
+    use crate::{InvoiceMutations, InvoiceQueries};
+
+    #[actix_rt::test]
+    async fn test_graphql_outbound_shipment_insert() {
+        let (mock_data, connection, _, settings) = setup_graphl_test(
+            InvoiceQueries,
+            InvoiceMutations,
+            "omsupply-database-gql-outbound_shipment_insert",
+            MockDataInserts::all(),
+        )
+        .await;
+
+        let other_party_supplier = &mock_data["base"].names[2];
+        let other_party_customer = &mock_data["base"].names[0];
+
+        let starting_invoice_number = mock_outbound_shipment_number_store_a().value;
+
+        let query = r#"mutation InsertOutboundShipment($input: InsertOutboundShipmentInput!) {
+            insertOutboundShipment(input: $input, storeId: \"store_a\") {
+                ... on InsertOutboundShipmentError {
+                  error {
+                    __typename
+                  }
+                }
+                ... on NodeError {
+                  error {
+                    __typename
+                  }
+                }
+                ... on InvoiceNode {
+                    id
+                    otherPartyId
+                    otherPartyStore {
+                      id
+                    }
+                    invoiceNumber
+                    type
+                    comment
+                    theirReference
+                    onHold
+                    colour
+                }
+            }
+        }"#;
+
+        // OtherPartyNotACustomerError
+        let variables = Some(json!({
+          "input": {
+            "id": "ci_insert_1",
+            "otherPartyId": other_party_supplier.id,
+          }
+        }));
+        let expected = json!({
+            "insertOutboundShipment": {
+              "error": {
+                "__typename": "OtherPartyNotACustomerError"
+              }
+            }
+          }
+        );
+        assert_graphql_query!(&settings, query, &variables, &expected, None);
+
+        // ForeignKeyError (OtherPartyIdNotFoundError)
+        let foreign_key_query = r#"mutation InsertOutboundShipment($input: InsertOutboundShipmentInput!) {
+          insertOutboundShipment(input: $input, storeId: \"store_a\") {
+             __typename
+          }
+        }"#;
+        let variables = Some(json!({
+          "input": {
+            "id": "ci_insert_1",
+            "otherPartyId": "not existing",
+          }
+        }));
+        let expected_message = "Bad user input";
+        assert_standard_graphql_error!(
+            &settings,
+            &foreign_key_query,
+            &variables,
+            &expected_message,
+            None,
+            None
+        );
+        // Test succeeding insert
+        let variables = Some(json!({
+          "input": {
+            "id": "ci_insert_1",
+            "otherPartyId": other_party_customer.id,
+            "comment": "ci comment",
+          }
+        }));
+        let expected = json!({
+            "insertOutboundShipment": {
+              "id": "ci_insert_1",
+              "invoiceNumber": starting_invoice_number+1,
+              "otherPartyId": other_party_customer.id,
+              "type": "OUTBOUND_SHIPMENT",
+              "comment": "ci comment",
+              "theirReference": null,
+              "onHold": false,
+              "colour": null,
+            }
+          }
+        );
+        assert_graphql_query!(&settings, query, &variables, &expected, None);
+        // make sure item has been inserted
+        InvoiceRepository::new(&connection)
+            .find_one_by_id("ci_insert_1")
+            .unwrap();
+
+        // Test succeeding insert on_hold and their_reference
+        let variables = Some(json!({
+          "input": {
+            "id": "ci_insert_2",
+            "otherPartyId": other_party_customer.id,
+            "theirReference": "reference",
+            "onHold": true,
+            "colour": "#FFFFFF"
+          }
+        }));
+        let expected = json!({
+            "insertOutboundShipment": {
+              "id": "ci_insert_2",
+              "invoiceNumber": starting_invoice_number+2,
+              "otherPartyId": other_party_customer.id,
+              "type": "OUTBOUND_SHIPMENT",
+              "comment": null,
+              "theirReference":"reference",
+              "onHold": true,
+              "colour": "#FFFFFF"
+            }
+          }
+        );
+        assert_graphql_query!(&settings, query, &variables, &expected, None);
+
+        // RecordAlreadyExist,
+        let variables = Some(json!({
+          "input": {
+            "id": "ci_insert_1",
+            "otherPartyId": other_party_customer.id,
+          }
+        }));
+        let expected_message = "Bad user input";
+        assert_standard_graphql_error!(
+            &settings,
+            &query,
+            &variables,
+            &expected_message,
+            None,
+            None
+        );
+
+        // Test Success name_store_id, linked to store
+        let id = uuid();
+        let variables = Some(json!({
+          "input": {
+            "id": id,
+            "otherPartyId": mock_name_linked_to_store().id,
+          }
+        }));
+        let expected = json!({
+            "insertOutboundShipment": {
+              "id": id,
+              "otherPartyStore": {
+                "id": mock_store_linked_to_name().id
+              }
+            }
+          }
+        );
+
+        assert_graphql_query!(&settings, query, &variables, &expected, None);
+
+        let new_invoice = InvoiceRepository::new(&connection)
+            .find_one_by_id(&id)
+            .unwrap();
+
+        assert_eq!(
+            new_invoice.name_store_id,
+            Some(mock_store_linked_to_name().id)
+        );
+
+        // Test Success name_store_id, not linked to store
+        let id = uuid();
+        let variables = Some(json!({
+          "input": {
+            "id": id,
+            "otherPartyId": mock_name_not_linked_to_store().id,
+          }
+        }));
+        let expected = json!({
+            "insertOutboundShipment": {
+              "id": id,
+              "otherPartyStore": null
+            }
+          }
+        );
+
+        assert_graphql_query!(&settings, query, &variables, &expected, None);
+
+        let new_invoice = InvoiceRepository::new(&connection)
+            .find_one_by_id(&id)
+            .unwrap();
+
+        assert_eq!(new_invoice.name_store_id, None);
+    }
+}
