+++ graphql/invoice/src/query_tests/mod.rs
@@ -1,0 +1,3 @@
+mod invoice_by_number;
+mod invoice_loaders;
+mod invoices;
