+++ graphql/src/lib.rs
@@ -10,7 +10,7 @@ use async_graphql::{EmptySubscription, SchemaBuilder};
 use async_graphql_actix_web::{Request, Response};
 use graphql_core::auth_data_from_request;
 use graphql_core::loader::LoaderRegistry;
-use graphql_invoice::InvoiceQueries;
+use graphql_invoice::{InvoiceMutations, InvoiceQueries};
 use mutations::Mutations;
 use queries::Queries;
 use repository::StorageConnectionManager;
@@ -20,13 +20,16 @@ use service::service_provider::ServiceProvider;
 #[derive(MergedObject, Default)]
 pub struct FullQuery(pub Queries, pub InvoiceQueries);
 
-pub type Schema = async_graphql::Schema<FullQuery, Mutations, async_graphql::EmptySubscription>;
-type Builder = SchemaBuilder<FullQuery, Mutations, EmptySubscription>;
+#[derive(MergedObject, Default)]
+pub struct FullMutation(pub Mutations, pub InvoiceMutations);
+
+pub type Schema = async_graphql::Schema<FullQuery, FullMutation, async_graphql::EmptySubscription>;
+type Builder = SchemaBuilder<FullQuery, FullMutation, EmptySubscription>;
 
 pub fn build_schema() -> Builder {
     Schema::build(
         FullQuery(Queries, InvoiceQueries),
-        Mutations,
+        FullMutation(Mutations, InvoiceMutations),
         EmptySubscription,
     )
 }
