+++ server/tests/graphql/mod.rs
@@ -11,15 +11,9 @@ use actix_web::web::Data;
 use serde_json::Value;
 
 mod common;
-// mod inbound_shipment_delete;
-// mod inbound_shipment_insert;
 // mod inbound_shipment_line_delete;
 // mod inbound_shipment_line_insert;
 // mod inbound_shipment_line_update;
-// mod inbound_shipment_update;
-mod invoice_by_number;
-mod invoice_loaders;
-mod invoices;
 mod item_stats;
 mod items;
 mod location_delete;
@@ -28,12 +22,9 @@ mod location_update;
 mod locations;
 mod master_lists;
 mod names;
-// mod outbound_shipment_delete;
-// mod outbound_shipment_insert;
 // mod outbound_shipment_line_delete;
 // mod outbound_shipment_line_insert;
 // mod outbound_shipment_line_update;
-// mod outbound_shipment_update;
 mod pagination;
 mod requisition;
 mod stocktake_batch;
